+++ fc-solve/source/fc_pro_iface_pos.h
@@ -37,7 +37,7 @@ void fc_solve_moves_processed_gen(fcs_moves_processed *,
     fcs_state_keyval_pair *, int, const fcs_moves_sequence_t *);
 void fc_solve_moves_processed_render_move(fcs_extended_move, char *);
 
-static inline int fc_solve_moves_processed_get_moves_left(
+static inline size_t fc_solve_moves_processed_get_moves_left(
     const fcs_moves_processed *const moves)
 {
     return moves->num_moves - moves->next_move_idx;
